cmd(profile): new flags logType and logFilter for profile command
